+++ Sources/MapboxSearch/InternalAPI/DefaultStringInterpolation+Extensions.swift
@@ -4,8 +4,13 @@ extension DefaultStringInterpolation {
         switch api {
         case .geocoding:
             versionName = "geocoding"
+
         case .SBS:
             versionName = "SBS"
+            
+        case .autofill:
+            versionName = "autofill"
+
         @unknown default:
             versionName = "unknown"
         }