+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Fixed
+- [Core] Fixed warning related to `CLLocationManager` permissions check.
+
 ## [1.0.0-beta.32] - 2022-07-04
 
 ### Added