+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 ### Added
 - [Core] `SearchResult.searchRequest` field exposed. Identifies original search request for a result.
+- [Core] `AddressAutofill` separate use case to for completing adress forms.
 - [Core] `SearchEngine.init(supportSBS: Bool) added flag to initialize `SearchEngine` with SBS support.
 
 ### Fixed